requirements.txt
fastapi
uvicorn
openai
pymupdf
python-multipart
python-dotenv
